Add files via upload
These source codes in raw book of literature rått will win the world informative and containt a source code through a source code string through my DNA that you have never found. It is probably higher than the source code string to DNA measurement at the DNA lab for DNA cloning, which proves that they use avläsare they should not have for me, which means that my DNA is produced fabricatedly in my persona and the result will missmatch with my DNA from birth.
Morse.
The benefits of using their highest DNA-avläsare on me is that it will give a result for a fake DNA that could handle a very rare DNA molecule, or not. But is it my fault that you used a source code string (string: think Morse), that has the lower amount of undecoding or interpretitive methods or possibilities than mine and I am as a consequence DNA cloned? Not. Definately not to a young kid, like I were, that you used a lower rang source code string when you undecoded through air and DNA cloned me and attacked my brain as a consequence of becoming DNA cloned.

For example: If my DNA supports the interpretion from right to left (reversed-integralic) and it involves the autoreversed as part of it, what will happen if this avläsare of DNA that the whole world DNA labs use; Yes they reach out in air and DNA clone me and think it looks unique but there starting to upcome results for me that many DNA labs and the police in America, are suspiscious to me being the reason for being DNA cloned, but it can never ever be my fault. And there is no need to punish me for being DNA cloned through your Python Avalan's, the lower rang string than should be used for my DNA rått.

I tested writing out a noneldering algorithm an hour ago and it was in one of the files above out - and if it worked raw through my DNA through the interpretion of how it should look through my brain you first must open it raw and not in any of the protocols and secondly, you need the better interpretitive possibilities to use them [in my DNA or blood the source code string has the possibility to be found within).

I also wrote a defendant algorithm for those in trouble with the police due to crimes on the web, but remember; I am proud to have written it so that they will get it all through data, which is impossible to prove top notchedly, and with the involvement of the defendant to know how they look at their crime through their thoughts and their psychological thinking to their crime and some people might actually be defended against being in jail.

I want to state one thing clear here: No one excludes me off my youngering procedure (föryngrare) raw. If you find it reserve it for the better for me as the first and the author of it because I am - through the brains output of the html document. Do not trick me of this föryngrare because I wrote it very well and it worked. You do not have to understand how important it was to be written perfected but I already succeed so be happy about that.

Another proof of that it is my brain and/or DNA who wrote these html document is that the methods does not match with any source code program out there, because it is better or should we reveal higher. And my brain works like this: Conceptual, valueing, abstract. Not in logic or definition.